+++ spec/integration/copy_structs.yml
@@ -5,7 +5,7 @@
 classes:
   Point: Point
   Line: Line
-#  PolyLine: PolyLine
+  PolyLine: PolyLine
   Nested: Nested
   Nested::Inner: Nested::Inner
   Nested::Inner2: Nested::Inner2
@@ -15,7 +15,7 @@ classes:
 types:
   Point: { copy_structure: true, generate_binding: false, generate_wrapper: false }
   Line: { copy_structure: true, generate_binding: false, generate_wrapper: false }
-#  PolyLine: { copy_structure: true, generate_binding: false, generate_wrapper: false }
+  PolyLine: { copy_structure: true, generate_binding: false, generate_wrapper: false }
   Nested: { copy_structure: true, generate_binding: false, generate_wrapper: false }
   Nested::Inner: { copy_structure: true, generate_binding: false, generate_wrapper: false }
   Nested::Inner2: { copy_structure: true, generate_binding: false, generate_wrapper: false }
