+++ src/bindgen/generator/crystal_lib.cr
@@ -33,13 +33,14 @@ module Bindgen
       def visit_struct(structure)
         puts "struct #{structure.name}"
         indented do
-          structure.fields.each do |name, type|
-            # can't use Void as a struct field type
-            if type.type_name == "Void"
-              puts "#{name} : #{type.type_name}*"
-            else
-              puts "#{name} : #{type.type_name}"
-            end
+          structure.fields.each do |name, result|
+            # can't use Void as a struct field type directly
+            ptr = result.pointer
+            ptr = {ptr, 1}.max if result.type_name == "Void"
+
+            stars = "*" * ptr if ptr > 0
+
+            puts "#{name} : #{result.type_name}#{stars}"
           end
         end
         puts "end"
